Move createOverrideWeight() into creditcoin-js v0.9.5
