+++ creditcoin-js/package.json
@@ -1,6 +1,6 @@
 {
     "name": "creditcoin-js",
-    "version": "0.9.0",
+    "version": "0.9.5",
     "author": "Jeremy Frank <jeremy.frank@gluwa.com>",
     "license": "MIT",
     "main": "lib/index.js",
