+++ creditcoin-js/src/utils.ts
@@ -3,6 +3,8 @@ import { joinSignature } from '@ethersproject/bytes';
 import { sha256AsU8a, blake2AsU8a } from '@polkadot/util-crypto';
 import { AccountId } from './model';
 import { ApiPromise } from '@polkadot/api';
+import { BN } from '@polkadot/util';
+import { SiLookupTypeId } from '@polkadot/types/interfaces/scaleInfo';
 
 export const signAccountId = (api: ApiPromise, signer: Wallet, accountId: AccountId) => {
     const accountIdBytes = api.createType('AccountId', accountId).toU8a();
@@ -10,8 +12,39 @@ export const signAccountId = (api: ApiPromise, signer: Wallet, accountId: Accoun
     return joinSignature(signer._signingKey().signDigest(accountIdHash)); // eslint-disable-line no-underscore-dangle
 };
 
+type OldWeight = BN;
+type NewWeight = { refTime: BN; proofSize: BN };
+type Weight = OldWeight | NewWeight;
+
+export const createOverrideWeight = (api: ApiPromise): Weight => {
+    const sudoCallTypeId = api.runtimeMetadata.registry.metadata.pallets
+        .find((v) => v.name.toString() === 'Sudo')
+        ?.calls.unwrap().type as SiLookupTypeId;
+    const sudoCallTypeDef = api.runtimeMetadata.registry.lookup.getTypeDef(sudoCallTypeId);
+    if (Array.isArray(sudoCallTypeDef.sub)) {
+        const sudoUncheckedWeightType = sudoCallTypeDef.sub.find((def) => def.name === 'sudo_unchecked_weight');
+        if (sudoUncheckedWeightType && Array.isArray(sudoUncheckedWeightType.sub)) {
+            // the weight is the second argument to sudo_unchecked_weight
+            const weightType = sudoUncheckedWeightType.sub[1];
+            if (weightType.type === 'u64') {
+                // old weight (simple u64)
+                return new BN(1);
+            } else {
+                // new weight
+                return {
+                    refTime: new BN(1),
+                    proofSize: new BN(0),
+                };
+            }
+        }
+    }
+
+    throw new Error("Couldn't find expected Weight type from sudoUncheckedWeight metadata");
+};
+
 export const utils = (api: ApiPromise) => {
     return {
         signAccountId: (signer: Wallet, accountId: AccountId) => signAccountId(api, signer, accountId),
+        createOverrideWeight: () => createOverrideWeight(api),
     };
 };
